feat: finish insert

+++ stage6/bufHash.C
@@ -10,50 +10,51 @@
 
 // buffer pool hash table implementation
 
-int BufHashTbl::hash(const File* file, const int pageNo)
+int BufHashTbl::hash(const File *file, const int pageNo)
 {
   long tmp, value;
-  tmp = (int)(long)file;  // cast of pointer to the file object to an integer
+  tmp = (long)file; // cast of pointer to the file object to an integer
   value = (tmp + pageNo) % HTSIZE;
   return value;
 }
-
 
 BufHashTbl::BufHashTbl(int htSize)
 {
   HTSIZE = htSize;
   // allocate an array of pointers to hashBuckets
-  ht = new hashBucket* [htSize];
-  for(int i=0; i < HTSIZE; i++)
+  ht = new hashBucket *[htSize];
+  for (int i = 0; i < HTSIZE; i++)
     ht[i] = NULL;
 }
 
-
 BufHashTbl::~BufHashTbl()
 {
-  for(int i = 0; i < HTSIZE; i++) {
-    hashBucket* tmpBuf = ht[i];
-    while (ht[i]) {
+  for (int i = 0; i < HTSIZE; i++)
+  {
+    hashBucket *tmpBuf = ht[i];
+    while (ht[i])
+    {
       tmpBuf = ht[i];
       ht[i] = ht[i]->next;
       delete tmpBuf;
     }
   }
-  delete [] ht;
+  delete[] ht;
 }
-
 
 //---------------------------------------------------------------
 // insert entry into hash table mapping (file,pageNo) to frameNo;
 // returns OK if OK, HASHTBLERROR if an error occurred
 //---------------------------------------------------------------
 
-Status BufHashTbl::insert(const File* file, const int pageNo, const int frameNo) {
+Status BufHashTbl::insert(const File *file, const int pageNo, const int frameNo)
+{
 
   int index = hash(file, pageNo);
 
-  hashBucket* tmpBuc = ht[index];
-  while (tmpBuc) {
+  hashBucket *tmpBuc = ht[index];
+  while (tmpBuc)
+  {
     if (tmpBuc->file == file && tmpBuc->pageNo == pageNo)
       return HASHTBLERROR;
     tmpBuc = tmpBuc->next;
@@ -62,7 +63,7 @@ Status BufHashTbl::insert(const File* file, const int pageNo, const int frameNo)
   tmpBuc = new hashBucket;
   if (!tmpBuc)
     return HASHTBLERROR;
-  tmpBuc->file = (File*) file;
+  tmpBuc->file = (File *)file;
   tmpBuc->pageNo = pageNo;
   tmpBuc->frameNo = frameNo;
   tmpBuc->next = ht[index];
@@ -71,18 +72,18 @@ Status BufHashTbl::insert(const File* file, const int pageNo, const int frameNo)
   return OK;
 }
 
-
-//-------------------------------------------------------------------	     
+//-------------------------------------------------------------------
 // Check if (file,pageNo) is currently in the buffer pool (ie. in
-// the hash table).  If so, return corresponding frameNo. else return 
+// the hash table).  If so, return corresponding frameNo. else return
 // HASHNOTFOUND
 //-------------------------------------------------------------------
 
-Status BufHashTbl::lookup(const File* file, const int pageNo, int& frameNo) 
-  {
+Status BufHashTbl::lookup(const File *file, const int pageNo, int &frameNo)
+{
   int index = hash(file, pageNo);
-  hashBucket* tmpBuc = ht[index];
-  while (tmpBuc) {
+  hashBucket *tmpBuc = ht[index];
+  while (tmpBuc)
+  {
     if (tmpBuc->file == file && tmpBuc->pageNo == pageNo)
     {
       frameNo = tmpBuc->frameNo; // return frameNo by reference
@@ -93,27 +94,31 @@ Status BufHashTbl::lookup(const File* file, const int pageNo, int& frameNo)
   return HASHNOTFOUND;
 }
 
-
 //-------------------------------------------------------------------
 // delete entry (file,pageNo) from hash table. REturn OK if page was
 // found.  Else return HASHTBLERROR
 //-------------------------------------------------------------------
 
-Status BufHashTbl::remove(const File* file, const int pageNo) {
+Status BufHashTbl::remove(const File *file, const int pageNo)
+{
 
   int index = hash(file, pageNo);
-  hashBucket* tmpBuc = ht[index];
-  hashBucket* prevBuc = ht[index];
+  hashBucket *tmpBuc = ht[index];
+  hashBucket *prevBuc = ht[index];
 
-  while (tmpBuc) {
-    if (tmpBuc->file == file && tmpBuc->pageNo == pageNo) {
-      if (tmpBuc == ht[index]) 
-	ht[index] = tmpBuc->next;
+  while (tmpBuc)
+  {
+    if (tmpBuc->file == file && tmpBuc->pageNo == pageNo)
+    {
+      if (tmpBuc == ht[index])
+        ht[index] = tmpBuc->next;
       else
-	prevBuc->next = tmpBuc->next;
+        prevBuc->next = tmpBuc->next;
       delete tmpBuc;
       return OK;
-    } else {
+    }
+    else
+    {
       prevBuc = tmpBuc;
       tmpBuc = tmpBuc->next;
     }
